+++ lib/low_type.rb
@@ -3,7 +3,9 @@
 require_relative 'adapters/adapter_loader'
 require_relative 'basic_types'
 require_relative 'instance_types'
+require_relative 'local_types'
 require_relative 'redefiner'
+require_relative 'subclasses'
 require_relative 'type_expression'
 require_relative 'value_expression'
 
@@ -11,11 +13,6 @@ require_relative 'value_expression'
 module LowType
   # We do as much as possible on class load rather than on instantiation to be thread-safe and efficient.
   def self.included(klass) # rubocop:disable Metrics/AbcSize
-    # Array[] and Hash[] class method returns a type expression only for the duration of this "included" hook.
-    array_class_method = Array.method('[]').unbind
-    hash_class_method = Hash.method('[]').unbind
-    LowType.redefine(hash_class_method:)
-
     class << klass
       def low_methods
         @low_methods ||= {}
@@ -27,6 +24,7 @@ module LowType
     private_start_line = parser.private_start_line
 
     klass.extend InstanceTypes
+    klass.include LocalTypes
     klass.prepend LowType::Redefiner.redefine(method_nodes: parser.instance_methods, klass:, private_start_line:, file_path:)
     klass.singleton_class.prepend LowType::Redefiner.redefine(method_nodes: parser.class_methods, klass:, private_start_line:, file_path:)
 
@@ -34,26 +32,18 @@ module LowType
       adapter.process
       klass.prepend Adapter::Methods
     end
-  ensure
-    Array.define_singleton_method('[]', array_class_method)
-    Hash.define_singleton_method('[]', hash_class_method)
   end
 
   class << self
     # Public API.
 
     def config
-      config = Struct.new(:local_types, :deep_type_check)
-      @config ||= config.new(false, false)
+      config = Struct.new(:deep_type_check, :severity_level)
+      @config ||= config.new(false, :error)
     end
 
     def configure
       yield(config)
-
-      return unless config.local_types
-
-      require_relative 'local_types'
-      include LocalTypes
     end
 
     # Internal API.
@@ -88,25 +78,6 @@ module LowType
 
     def value(type:)
       TypeExpression.new(default_value: ValueExpression.new(value: type))
-    end
-
-    # TODO: Unit test.
-    def redefine(hash_class_method:)
-      Array.define_singleton_method('[]') do |*types|
-        TypeExpression.new(type: [*types])
-      end
-
-      Hash.define_singleton_method('[]') do |type|
-        # Support Pry which uses Hash[].
-        unless LowType.type?(type)
-          Hash.define_singleton_method('[]', hash_class_method)
-          result = Hash[type]
-          Hash.method('[]').unbind
-          return result
-        end
-
-        TypeExpression.new(type:)
-      end
     end
   end
 end