+++ lib/type_expression.rb
@@ -56,7 +56,15 @@ module LowType
     end
 
     def valid_types
-      types = @types.map { |type| type.inspect.to_s }
+      types = @types.map do |type|
+        # Remove 'LowType::' namespaces in subtypes.
+        if type.is_a?(::Array)
+          '[' + type.map { |subtype| subtype.to_s.delete_prefix('LowType::') }.join(', ') + ']'
+        else
+          type.inspect.to_s.delete_prefix('LowType::')
+        end
+      end
+
       types += ['nil'] if @default_value.nil?
       types.join(' | ')
     end
@@ -94,26 +102,6 @@ module LowType
       proxy_file_backtrace = "#{from_prefix}#{proxy_file_backtrace}" if from_prefix
 
       [proxy_file_backtrace, *filtered_backtrace]
-    end
-  end
-end
-
-# For "Type | [type_expression/type/value]" situations, redirecting to or generating a type expression from types.
-# "|" is not defined on Object class and this is the most compute-efficient way to achieve our goal (world peace).
-# "|" is overridable by any child object. While we could def/undef this method, this approach is actually lighter.
-# "|" bitwise operator on Integer is not defined when the receiver is an Integer class, so we are not in conflict.
-class Object
-  class << self
-    def |(expression)
-      if expression.instance_of?(::LowType::TypeExpression)
-        # We pass our type into their type expression.
-        expression | self
-        expression
-      else
-        # We turn our type into a type expression and pass in their [type_expression/type/value].
-        type_expression = ::LowType::TypeExpression.new(type: self)
-        type_expression | expression
-      end
     end
   end
 end